+++ tests/test_tensor.py
@@ -56,8 +56,8 @@ def test_matmul(interpret, with_bounds_inference):
         return array_(lambda i, j: sum_(lambda t: a[i, t] * b[t, j]))
 
     (a0, b0), matmul_expr = function([matrix(float), matrix(float)], matmul)
-    first = numpy.array([[1, 2, 3], [4, 5, 6]])
-    second = numpy.array([[1], [0], [-1]])
+    first = numpy.array([[1, 2, 3], [4, 5, 6]], dtype=float)
+    second = numpy.array([[1], [0], [-1]], dtype=float)
     numpy.testing.assert_allclose(
         interpret(
             matmul_expr,
